# src/main/java/com/swifttech/edx/sr/ps/adapter/BusinessEntityCategoryController.java
package com.swifttech.edx.sr.ps.adapter;

import com.swifttech.edx.dm.exception.GlobalException;
import com.swifttech.edx.dm.payload.response.GlobalResponse;
import com.swifttech.edx.sr.ps.annotation.ProductAndServicesRestController;
import com.swifttech.edx.sr.ps.model.request.BusinessEntityCategoryCreateUpdateRequest;
import com.swifttech.edx.sr.ps.model.request.BusinessEntityCategoryDataRequest;
import com.swifttech.edx.sr.ps.service.BusinessEntityCategoryService;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;

@ProductAndServicesRestController
@RequiredArgsConstructor
public class BusinessEntityCategoryController {

    private final BusinessEntityCategoryService businessEntityCategoryService;

    @PostMapping("business-entity-category/create")
    public ResponseEntity<GlobalResponse> create(@RequestBody @Valid BusinessEntityCategoryCreateUpdateRequest request) throws GlobalException {
        return ResponseEntity.ok(businessEntityCategoryService.createBusinessEntityCategory(request));
    }

    @PutMapping("business-entity-category/{id}/update")
    public ResponseEntity<GlobalResponse> update(@PathVariable Long id, @RequestBody @Valid BusinessEntityCategoryCreateUpdateRequest request) throws GlobalException {
        return ResponseEntity.ok(businessEntityCategoryService.updateBusinessEntityCategory(id, request));
    }

    @GetMapping("business-entity-category/{id}")
    public ResponseEntity<GlobalResponse> findById(@PathVariable Long id) throws GlobalException {
        return ResponseEntity.ok(businessEntityCategoryService.findBusinessEntityCategoryById(id));
    }

    @PostMapping("business-entity-category/list")
    public ResponseEntity<GlobalResponse> findAllPaginated(@RequestBody BusinessEntityCategoryDataRequest request) throws GlobalException {
        return ResponseEntity.ok(businessEntityCategoryService.findBusinessEntityCategoryPaginatedData(request));
    }

}